# docs/tecnicos/20260121a-openapi_copiloto_v2.yaml
openapi: 3.0.3
info:
  title: Copiloto de Emprendimiento v2.0 API
  description: |
    API del sistema de asistencia inteligente para validación de modelos de negocio.
    
    ## Autenticación
    Todas las peticiones requieren autenticación mediante JWT token en el header `Authorization: Bearer {token}`.
    
    ## Rate Limiting
    - Chat API: 60 peticiones/minuto por usuario
    - Otros endpoints: 100 peticiones/minuto por usuario
    
    ## Errores
    Los errores siguen el formato estándar JSON:API con códigos HTTP apropiados.
  version: 2.0.0
  contact:
    name: Jaraba Impact Platform
    email: tech@jaraba.io
  license:
    name: Proprietary
servers:
  - url: https://api.jaraba.io/v2
    description: Producción
  - url: https://staging-api.jaraba.io/v2
    description: Staging
  - url: http://localhost:8080/v2
    description: Desarrollo local
...
paths:
...
  /copilot/chat:
    post:
      tags:
        - Copilot
      summary: Enviar mensaje al Copiloto
      description: |
        Envía un mensaje al Copiloto y recibe una respuesta personalizada.
        El sistema automáticamente:
        - Carga el contexto del emprendedor
        - Detecta el modo apropiado (Coach, Consultor, Sparring, CFO, Abogado)
        - Detecta emociones si las hay
        - Sugiere experimentos relevantes
      operationId: chatWithCopilot
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChatRequest'
            example:
              user_id: "550e8400-e29b-41d4-a716-446655440000"
              session_id: "session-abc-123"
              message: "No sé si debería cobrar 50€ por hora, me parece caro"
      responses:
        '200':
          description: Respuesta exitosa del Copiloto
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatResponse'
              example:
                response: "Entiendo perfectamente esa sensación, María. Es muy común cuando empezamos..."
                mode_detected: "CFO_SINTETICO"
                emotion_detected: "miedo_precio"
                suggested_actions:
                  - type: "TOOL"
                    label: "Abrir Calculadora de la Verdad"
                    url: "/tools/calculadora-verdad"
                experiment_suggested:
                  id: 28
                  name: "Test A/B de Precio"
                  reason: "Para validar si el mercado acepta ese precio"
                tokens_used: 847
                session_id: "session-abc-123"
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/RateLimitExceeded'
...
components:
  schemas:
...
    ChatRequest:
      type: object
      required:
        - user_id
        - message
      properties:
        user_id:
          type: string
          format: uuid
        session_id:
          type: string
        message:
          type: string
          maxLength: 4000
    ChatResponse:
      type: object
      properties:
        response:
          type: string
        mode_detected:
          type: string
          enum: [COACH_EMOCIONAL, CONSULTOR_TACTICO, SPARRING_PARTNER, CFO_SINTETICO, ABOGADO_DIABLO]
        emotion_detected:
          type: string
          nullable: true
        suggested_actions:
          type: array
          items:
            type: object
            properties:
              type:
                type: string
                enum: [TOOL, EXPERIMENT, HYPOTHESIS, LINK]
              label:
                type: string
              url:
                type: string
        experiment_suggested:
          type: object
          nullable: true
          properties:
            id:
              type: integer
            name:
              type: string
            reason:
              type: string
        tokens_used:
          type: integer
        session_id:
          type: string
...
  responses:
    BadRequest:
      description: Petición inválida
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
              details:
                type: array
                items:
                  type: string
    Unauthorized:
      description: No autenticado
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: "Token inválido o expirado"
...
    RateLimitExceeded:
      description: Límite de peticiones excedido
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: "Has excedido el límite de peticiones. Espera 60 segundos."
              retry_after:
                type: integer
                example: 60
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []